Add the automated exploit generation toolset.
This commit adds a GDB script that dumps EBP and BUFFPTR to the
file 'offset.txt', as well as new Makefile targets.

Changes have been made to exploit.c so that it reads in two lines,
one hexadecimal address on each (for EBP and BUFFPTR, respectively).

A new gitignore was added so that offset.txt could be ignored.